cloud_functions/sync_drive_to_gcs/requirements.txt
functions-framework==3.*
google-cloud-storage==2.19.0
google-cloud-bigquery==3.30.0
google-api-python-client==2.150.0
google-auth==2.38.0
